Clean Code - Move 'GeometryAdapter' from api.model to api.mapping.
